create invoices and save image locally

+++ src/controllers/admin.js
@@ -1,5 +1,7 @@
+import get from 'lodash.get';
 import Product from '../models/Product';
 import { validationResult } from 'express-validator/check';
+import { deleteFile } from '../utils/fileHelpers';
 import handleServerError from '../middlewares/handleServerError';
 
 export function getAddProduct(req, res) {
@@ -14,14 +16,15 @@ export function getAddProduct(req, res) {
 
 export function postAddProduct(req, res) {
     const { _id: userId } = req.user;
-    const { title, price, imageUrl, description } = req.body;
+    const { title, price, description } = req.body;
+    const imageUrl = get(req.file, 'path');
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
         return res.status(422).render('layout', {
             route: 'admin_product_form',
             title: 'Add Product',
-            product: { title, price, imageUrl, description },
+            product: { title, price, description },
             validation: errors.array().map(({ params }) => params),
             errors: errors.array(),
         });
@@ -74,14 +77,15 @@ export function getEditProduct(req, res) {
 export function postEditProduct(req, res) {
     const { _id } = req.user;
     const { productId } = req.params;
-    const { title, price, imageUrl, description } = req.body;
+    const { title, price, description } = req.body;
+    const imageUrl = get(req.file, 'path');
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
         return res.status(422).render('layout', {
             route: 'admin_product_form',
-            title: `Edit Product - ${product.title}`,
-            product: { _id, title, price, imageUrl, description },
+            title: 'Edit Product',
+            product: { _id, title, price, description },
             validation: errors.array().map(({ params }) => params),
             errors: errors.array(),
         });
@@ -92,9 +96,12 @@ export function postEditProduct(req, res) {
             if (product.userId.toString() !== _id.toString()) {
                 return res.redirect('/');
             }
+            if (imageUrl) {
+                deleteFile(imageUrl);
+                product.imageUrl = imageUrl;
+            }
             product.title = title;
             product.price = price;
-            product.imageUrl = imageUrl;
             product.description = description;
             return product.save();
         })
@@ -106,7 +113,14 @@ export function postDeleteProduct(req, res) {
     const { _id: userId } = req.user;
     const { productId } = req.body;
 
-    Product.deleteOne({ _id: productId, userId })
+    Product.findById(productId)
+        .then(product => {
+            if (!product) {
+                return handleServerError('Product not found');
+            }
+            deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: productId, userId });
+        })
         .then(() => res.redirect('/admin/products'))
         .catch(handleServerError);
 }
